requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
stripe==7.4.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
google-generativeai==0.8.3
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
reportlab==4.0.7
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
markdown==3.5.1

